virtualizor: router/profile refactoring
Do not use anymore the profile of the host to know if we should add
a NATed NIC on the public_network.
Instead, we now have to configure a NIC on the machine with the folowing
keys:
 -
   nat: True
   network_name: __public_network_

It's also possible to configure the `bootproto` to get the IP from a
DHCP server. But the use of a DHCP is not mandatoried anymore.